% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_helpers.R
\docType{data}
\name{co_colours}
\alias{co_colours}
\title{Vector of CO colours}
\format{
An object of class \code{character} of length 11.
}
\usage{
co_colours
}
\description{
A vector of Cabinet Office's corporate palette
}
\keyword{datasets}
