% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_deck.R
\name{co_deck}
\alias{co_deck}
\title{CO data reports}
\usage{
co_deck(css = NULL, ...)
}
\arguments{
\item{css}{Additional css files (default is \code{NULL})}

\item{...}{Arguments to pass on to \code{\link[pagedown:html_paged]{pagedown::html_paged()}}}
}
\value{

}
\description{
CO data reports
}
