% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{co_logo}
\alias{co_logo}
\title{CO logo}
\usage{
co_logo(out = c("html", "path"))
}
\arguments{
\item{out}{Either \code{html} or \code{path}}
}
\value{

}
\description{
A simple function to return the embedded logo, either as an html element or
the file path to the image.
}
