% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_svg.R
\name{export_plot}
\alias{export_plot}
\title{Export report plot as PNG}
\usage{
export_plot(gg_plot, filename, width, height)
}
\arguments{
\item{gg_plot}{The ggplot to render}

\item{filename}{The file to export}

\item{width}{The width in report columns}

\item{height}{The height in report columns}
}
\description{
Export report plots with a sizing similar to that of the plot created by
\code{\link[=insert_svg]{insert_svg()}}. Plots are exported as SVG if the file extension is \code{.svg}
otherwise they are exported in PNG format.
}
