% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_svg.R
\name{insert_svg}
\alias{insert_svg}
\title{Insert plot as SVG}
\usage{
insert_svg(gg_plot, width, height)
}
\arguments{
\item{gg_plot}{The ggplot to convert}

\item{width}{The width of the plot, in page columns}

\item{height}{The height of the plot, in page columns}
}
\value{

}
\description{
A function to convert a ggplot object to an SVG image for rendering
in the deckhand column/row grid structure
}
