% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\name{output_table}
\alias{output_table}
\title{Export a {gt} pipeline into output format}
\usage{
output_table(
  x,
  id = NULL,
  min_max_styling = FALSE,
  drop_header = FALSE,
  small = FALSE
)
}
\arguments{
\item{x}{The gt table}

\item{id}{A unique id for the table}

\item{min_max_styling}{Convert min/max flags to css classes}

\item{drop_header}{Drop the table header row(s)}

\item{small}{Whether to render in a smaller font}
}
\description{
Take an table constructed via the {gt} package and format it so
that it will work with the {pagedown} report template and css files.
}
