% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_helpers.R
\name{scale_co}
\alias{scale_colour_co}
\alias{scale_fill_co}
\title{Use CO colours for colour scales}
\usage{
scale_colour_co(..., colour, direction = 1, show.na = FALSE)

scale_fill_co(..., fill, direction = 1, show.na = FALSE)
}
\arguments{
\item{...}{Other arguments passed to discrete scale}

\item{colour}{A colour name from \code{\link[=co_colours]{co_colours()}} for scale_colour_co()}

\item{direction}{The direction of the scale}

\item{show.na}{Whether to show NA values}

\item{fill}{A colour name from \code{\link[=co_colours]{co_colours()}} for scale_fill_co()}
}
\description{
The scale_*_co() functions provide a way to easily apply a colour from the
CO corporate palette to a chart, the take a single colour name and apply
tints of that colour to the relevant series (colour/fill).
}
