% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_example_report}
\alias{show_example_report}
\title{Open example report in the browser}
\usage{
show_example_report(type = c("html", "pdf"), location = c("web", "local"))
}
\arguments{
\item{type}{Either \code{html} (the default) or \code{pdf}}

\item{location}{Either \code{web} (the default) or \code{local}}
}
\description{
View the example report
}
\examples{
\dontrun{
show_example_report()
}
}
