% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_helpers.R
\name{theme_co_report}
\alias{theme_co_report}
\title{Report plot theme}
\usage{
theme_co_report(
  base = "minimal",
  legend_position = "top",
  legend_title = "none",
  axis_text = "both",
  axis_title = "none"
)
}
\arguments{
\item{base}{The base ggplot theme (either minimal or void)}

\item{legend_position}{The location of the plot legend}

\item{legend_title}{The title for the legend or "none" (the default) for no
title}

\item{axis_text}{Which axis to include text for: "both" (the default), "x",
"y" or "none"}

\item{axis_title}{Which axis to include titles for: "both", "x", "y" or
"none" (the default)}
}
\description{
Provide common styling to ggplot objects
}
