% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_helpers.R
\name{tint_hex}
\alias{tint_hex}
\title{Produce tints of hex colours}
\usage{
tint_hex(colour, tint, bg_colour = "#ffffff")
}
\arguments{
\item{colour}{The colour to tint}

\item{tint}{The percentage tint desired, higher is 'lighter'}

\item{bg_colour}{The background colour to tint against, defaults to white}
}
\value{

}
\description{
A function to produce a tinted version of a hexadecimal colour code
}
