% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_example_report}
\alias{show_example_report}
\title{Open example report in the browser}
\usage{
show_example_report()
}
\description{
An example report showing layout options is provided in the package. This
function loads the report in your browser
}
